update readme, minor refactoring
