+++ 2-Authorization-I/1-call-graph/AppCreationScripts/sample.json
@@ -12,7 +12,7 @@
           "Name": "ms-identity-react-c2s1",
           "Kind": "SinglePageApplication",
           "HomePage": "http://localhost:3000/",
-          "ReplyUrls": "http://localhost:3000/,http://localhost:3000/redirect",
+          "ReplyUrls": "http://localhost:3000/, http://localhost:3000/redirect",
           "Audience": "AzureADMyOrg",
           "RequiredResourcesAccess": [
             {
@@ -42,14 +42,6 @@
                 {
                     "key": "Enter_the_Tenant_Info_Here",
                     "value": "$tenantId"
-                },
-                {
-                    "key": "Enter_the_Redirect_Uri",
-                    "value": "$spaAadApplication.Spa.RedirectUris[1]"
-                },
-                {
-                  "key": "Enter_the_Post_Redirect_Uri",
-                  "value": "$spaAadApplication.Spa.RedirectUris[0]"
                 }
             ]
         }
