+++ 2-Authorization-I/1-call-graph/SPA/package.json
@@ -18,6 +18,9 @@
   "devDependencies": {
     "react-scripts": "^5.0.1"
   },
+  "overrides": {
+    "autoprefixer": "10.4.5"
+  },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
