+++ 2-Authorization-I/1-call-graph/SPA/src/App.jsx
@@ -1,5 +1,4 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-
 import { MsalProvider } from '@azure/msal-react';
 
 import { PageLayout } from './components/PageLayout';
