+++ 2-Authorization-I/1-call-graph/SPA/src/authConfig.js
@@ -6,17 +6,17 @@
 import { LogLevel } from "@azure/msal-browser";
 
 /**
- * Configuration object to be passed to MSAL instance on creation. 
+ * Configuration object to be passed to MSAL instance on creation.
  * For a full list of MSAL.js configuration parameters, visit:
- * https://github.com/AzureAD/microsoft-authentication-library-for-js/blob/dev/lib/msal-browser/docs/configuration.md 
+ * https://github.com/AzureAD/microsoft-authentication-library-for-js/blob/dev/lib/msal-browser/docs/configuration.md
  */
 export const msalConfig = {
     auth: {
         clientId: 'Enter_the_Application_Id_Here', // This is the ONLY mandatory field that you need to supply.
         authority: 'https://login.microsoftonline.com/Enter_the_Tenant_Info_Here', // Defaults to "https://login.microsoftonline.com/common"
-        redirectUri: 'Enter_the_Redirect_Uri', // Points to window.location.origin. You must register this URI on Azure Portal/App Registration.
-        postLogoutRedirectUri: 'Enter_the_Post_Redirect_Uri', // Indicates the page to navigate after logout.
-        ClientCapabilities: ['cp1'], // this lets the resource owner know that this client is capable of handling claims challenge.
+        redirectUri: '/redirect', // Points to window.location.origin. You must register this URI on Azure Portal/App Registration.
+        postLogoutRedirectUri: '/', // Indicates the page to navigate after logout.
+        ClientCapabilities: ['CP1'], // this lets the resource owner know that this client is capable of handling claims challenge.
     },
     cache: {
         cacheLocation: 'localStorage', // Configures cache location. "sessionStorage" is more secure, but "localStorage" gives you SSO between tabs.
@@ -50,7 +50,7 @@ export const msalConfig = {
 /**
  * Scopes you add here will be prompted for user consent during sign-in.
  * By default, MSAL.js will add OIDC scopes (openid, profile, email) to any login request.
- * For more information about OIDC scopes, visit: 
+ * For more information about OIDC scopes, visit:
  * https://docs.microsoft.com/en-us/azure/active-directory/develop/v2-permissions-and-consent#openid-connect-scopes
  */
 export const loginRequest = {
@@ -65,8 +65,7 @@ export const protectedResources = {
     graphMe: {
         endpoint: "https://graph.microsoft.com/v1.0/me",
         scopes: ["User.Read"],
-    },
-    graphMessages: {
+    }, graphMessages: {
         endpoint: "https://graph.microsoft.com/v1.0/me/messages",
         scopes: ["Mail.Read"],
     },
