+++ 2-Authorization-I/1-call-graph/SPA/src/index.js
@@ -23,9 +23,8 @@ if (!msalInstance.getActiveAccount() && msalInstance.getAllAccounts().length > 0
 // Optional - This will update account state if a user signs in from another tab or window
 msalInstance.enableAccountStorageEvents();
 
-
 /**
-* To set an active account after the user signs in, register an event and listen to LOGIN_SUCCESS. For more, 
+* To set an active account after the user signs in, register an event and listen to LOGIN_SUCCESS. For more,
 * visit: https://github.com/AzureAD/microsoft-authentication-library-for-js/blob/dev/lib/msal-react/docs/events.md
 */
 msalInstance.addEventCallback((event) => {
