+++ 2-Authorization-I/1-call-graph/SPA/src/pages/Mails.jsx
@@ -1,11 +1,13 @@
 import { useEffect, useState } from 'react';
 import { MsalAuthenticationTemplate } from '@azure/msal-react';
 import { InteractionType } from '@azure/msal-browser';
+
 import { loginRequest } from '../authConfig';
 import { MailsData } from '../components/DataDisplay';
 import { protectedResources } from '../authConfig';
-import useTokenAcquisition from '../customHooks/useTokenAcquisition';
 import { getGraphClient } from '../graph';
+
+import useTokenAcquisition from '../hooks/useTokenAcquisition';
 
 const MailsContent = () => {
     const [response] = useTokenAcquisition(protectedResources.graphMessages.scopes);
@@ -42,7 +44,10 @@ export const Mails = () => {
     };
 
     return (
-        <MsalAuthenticationTemplate interactionType={InteractionType.Redirect} authenticationRequest={authRequest}>
+        <MsalAuthenticationTemplate
+            interactionType={InteractionType.Redirect}
+            authenticationRequest={authRequest}
+        >
             <MailsContent />
         </MsalAuthenticationTemplate>
     );
