--- 2-Authorization-I/1-call-graph/SPA/src/styles/AccountPicker.module.css
@@ -1,3 +1,0 @@
-.iconText {
-    margin: 0 .5rem
-}
