+++ 2-Authorization-I/1-call-graph/SPA/src/styles/App.css
@@ -9,3 +9,15 @@
   justify-content: center;
   text-align: left;
 }
+
+.iconText {
+  margin: 0 .5rem
+}
+
+.navbarStyle {
+  padding: .5rem 1rem
+}
+
+.navbarButton {
+  color: #fff !important
+}
