--- 2-Authorization-I/1-call-graph/SPA/src/styles/Navigation.module.css
@@ -1,7 +1,0 @@
-.navbarStyle {
-    padding: .5rem 1rem
-}
-
-.navbarButton {
-    color: #fff !important
-}
